Added point shifting with leading '+' or trailing '-'
Quick implementation, relies on specific formatting for ASCII lines, must lead with a space and not have an ending space so the characters are at an even number.  Will crash otherwise.

Points can be shifted forward with a + before the point and shift backward with a - after the point. '·+3 1 5-·' would all have the same starting point.